# Improve .extend() performance